Asset Bundles Are now supported through BundleLoadingInfo
