+++ MonsterTrainModdingAPI/Utilities/AssetBundleUtils.cs
@@ -8,19 +8,6 @@ using System.Reflection;
 
 namespace MonsterTrainModdingAPI.Utilities
 {
-    public struct AssetBundleLoadingInfo
-    {
-        public string bundleName;
-        public string assetName;
-        public string pathExtension;
-        public AssetBundleLoadingInfo(string bundleName, string assetName, string pathExtension = "")
-        {
-            this.assetName = assetName;
-            this.bundleName = bundleName;
-            this.pathExtension = pathExtension;
-        }
-    }
-
     public class AssetBundleUtils
     {
         private static Dictionary<string, AssetBundle> LoadedAssetBundles = new Dictionary<string, AssetBundle>();
